Fixed GUI bug.
