+++ gits/ui/lists_display/bands/band_list_widget.py
